Fixed extracting a zip file issue
Even though you select to unzip your file upload
it was uploading the zip file and not it's content.
The bug is in the check for zip extension. It was
checking for "zip" instead of ".zip". Path.GetExtension
includes the ".".

See
"http://msdn.microsoft.com/en-us/library/system.io.path.getextension.aspx"