Updated Disqus to use the web API
Updated Disqus to use the web API and changed the way we get the number
of comments from Disqus. The display of the comment count was changed so
we have a uniform CSS styling for DB comments, Disqus comments and when
we have the comments disabled.
We have now a new field on the admin page to enter our Disqus API Key