Theme handling rework
Theme CSS files are now loaded as bundles (file lists are generated at
startup for each theme).

The images files are moved back next to the CSS files into the \Themes\
folder (and they are finally displayed correctly).

Themes are applied to the administrator area as well.

Aplha version of the support of Bootstrap and WordPress themes are
implemented.